--- app/src/main/java/com/hellmund/primetime/ui/shared/SingleLiveDataEvent.kt
@@ -1,18 +1,0 @@
-package com.hellmund.primetime.ui.shared
-
-open class SingleLiveDataEvent<T>(
-    private val value: T
-) {
-
-    private var hasBeenHandled = false
-
-    fun getIfNotHandled(): T? {
-        return if (hasBeenHandled) {
-            null
-        } else {
-            hasBeenHandled = true
-            value
-        }
-    }
-
-}
